Speckart App to buy, manage orders
Specskart is a project created to show the functionalities of a Specs booking app.
The app has option to create new user and user login options. Options to add product to cart and checkout. Creating a eye checkup appointment.
Created on angular for the front end and backend on Spring Boot.